Fix phone layout overflow
- Let the main column and grid cells shrink so wide tables scroll inside
  their cards instead of stretching every page off-screen on phones
- Keep data-table rows on one line on phones, with edge scroll hints
- Hide keyboard-only top-bar items on phones and enlarge touch targets
- Round batch day counts (Carbon 3 returns floats), wrap the grade legend
  on narrow cards, and only show the mobile sidebar shadow while open

+++ resources/views/batches/show.blade.php
@@ -11,9 +11,9 @@
     $gradeB       = $records->where('quality_grade','B')->sum('quantity_kg');
     $gradeC       = $records->where('quality_grade','C')->sum('quantity_kg');
     $flushCount   = $records->count();
-    $daysSince    = $batch->inoculation_date->diffInDays(now());
+    $daysSince    = (int) $batch->inoculation_date->diffInDays(now());
     $daysLeft     = now()->lt($batch->expected_harvest_date)
-                        ? now()->diffInDays($batch->expected_harvest_date)
+                        ? (int) now()->diffInDays($batch->expected_harvest_date)
                         : null;
     $isOverdue    = now()->gt($batch->expected_harvest_date) && !in_array($batch->status, ['harvested','completed','contaminated']);
     $statusMap    = ['planned'=>'badge-gray','inoculated'=>'badge-blue','fruiting'=>'badge-green','harvested'=>'badge-yellow','completed'=>'badge-green','contaminated'=>'badge-red'];
